+++ Frends.MongoDB.Index/Frends.MongoDB.Index/Definitions/Enums.cs
@@ -1,0 +1,16 @@
+﻿namespace Frends.MongoDB.Index.Definitions;
+
+/// <summary>
+/// Index options.
+/// </summary>
+public enum IndexAction
+{
+	/// <summary>
+	/// Crrates an index.
+	/// </summary>
+	Create,
+	/// <summary>
+	/// Deltes an index.
+	/// </summary>
+	Drop
+}