+++ Frends.MongoDB.Query/Frends.MongoDB.Query/Definitions/Input.cs
@@ -10,9 +10,9 @@ public class Input
 	/// <summary>
 	/// Query single or multiple documents.
 	/// </summary>
-	/// <example>QueryOptions.QueryMany</example>
+	/// <example>DeleteOptions.DeleteOne</example>
 	[DefaultValue(QueryOptions.QueryMany)]
-	public QueryOptions QueryOptions { get; set; } = QueryOptions.QueryMany;
+	public QueryOptions QueryOptions { get; set; }
 
 
 	/// <summary>
@@ -20,4 +20,11 @@ public class Input
 	/// </summary>
 	/// <example>{'foo':'bar'}</example>
 	public string Filter { get; set; }
+
+	/// <summary>
+	/// Json output mode.
+	/// </summary>
+	/// <example>JsonOutputMode.RelaxedExtendedJson</example>
+	[DefaultValue(JsonOutputMode.RelaxedExtendedJson)]
+	public JsonOutputMode JsonOutputMode { get; set; }
 }